Update invocation of command to drop database
Newer versions of mongo (such as the one currently in use in our labs, 7) use `mongosh` rather than just `mongo` to do this step. This was leading to the database not being dropped in the script, which was confusing since there is output that suggests it was meant to happen.