Update GitHub settings and workflows
I mostly updated these to match what we have in the iteration template. The one exception is that many checks are run in the template on both push and pull request, and I think I'd like to try what was in Lab 4 previously and just do checks on push. I think some things were being run twice on PRs.