Update test to remove call to `input()`
This test contained an unsupported call to the `input` function (according to an error (maybe from eslint, but not sure about that)). I think this is a result of the switch to signals, but maybe not. Here is a link to what I used as a guide: https://angular.dev/guide/testing/components-scenarios#test-dashboardherocomponent-standalone