Update files according to some eslint rules
I added some rules where I would be able to see the corrections and feel more confident that eslint was working. It's not generally the goal of eslint to fix formatting and focus on what the code looks like. I'm going to keep the changes to the rules on a separate PR, but I did make these changes based on the rules. It does make things more consistent, even if it's mostly cosmetic.

        // The following few rules are more about how the code looks than how it functions,
        // but these are places where consistency stands out and can be confusing.
        // angular-eslint is more focused on functionality, so this is not promised to work forever:
        // https://github.com/angular-eslint/angular-eslint/blob/main/docs/FORMATTING_RULES.md
        "indent": ["error", 2],         // https://archive.eslint.org/docs/rules/indent
        "space-before-blocks": "error", // https://archive.eslint.org/docs/rules/space-before-blocks
        "no-trailing-spaces": "error",  // https://archive.eslint.org/docs/rules/no-trailing-spaces
        "brace-style": "error",         // https://archive.eslint.org/docs/rules/brace-style
        "func-call-spacing": "error",   // https://archive.eslint.org/docs/rules/func-call-spacing
        "arrow-spacing": "error",       // https://archive.eslint.org/docs/rules/arrow-spacing