Try another way to test without changing to async
I didn't realize I could make fewer changes to the code to use setInput rather than input. I was previously making this asynchronous instead of running it in the injection context. This commit makes the code look a lot more like the test looked when it was still using input. I'm not sure which approach is "better".